+++ src/main.rs
@@ -1,11 +1,14 @@
+use askama::Template;
 use axum::{
     body,
+    handler::Handler,
     http::header::CONTENT_TYPE,
     http::{Response, StatusCode},
     response::IntoResponse,
     routing::{get, get_service},
     Router,
 };
+use handlers::HtmlTemplate;
 use std::net::SocketAddr;
 
 use prometheus::{Encoder, TextEncoder};
@@ -17,10 +20,7 @@ use crate::handlers::{
     series::{series_handler, series_index_handler},
 };
 
-use http_body::{Body as _, Full};
-use std::io;
-use tower::ServiceBuilder;
-use tower_http::services::fs::{ServeDir, ServeFileSystemResponseBody};
+use tower_http::services::{fs::ServeDir, ServeFile};
 
 mod errors;
 mod handlers;
@@ -28,29 +28,22 @@ mod parsers;
 
 #[tokio::main]
 async fn main() {
-    let series_routes = Router::new()
-        .route("/", get(series_index_handler))
-        .route("/:series", get(series_handler));
-
-    let handler_404 = ServiceBuilder::new().and_then(
-        |response: Response<ServeFileSystemResponseBody>| async move {
-            let response = if response.status() == StatusCode::NOT_FOUND {
-                let body = Full::from("Not Found").map_err(|err| match err {}).boxed();
-                Response::builder()
-                    .status(StatusCode::NOT_FOUND)
-                    .body(body)
-                    .unwrap()
-            } else {
-                response.map(|body| body.boxed())
-            };
-
-            Ok::<_, io::Error>(response)
-        },
-    );
-
     let app = Router::new()
-        .fallback(
-            get_service(handler_404.service(ServeDir::new("./public"))).handle_error(
+        .fallback(handler_404.into_service())
+        .route(
+            "/sw.js",
+            get_service(ServeFile::new("./public/js/sw.js")).handle_error(
+                |error: std::io::Error| async move {
+                    (
+                        StatusCode::INTERNAL_SERVER_ERROR,
+                        format!("Unhandled internal error: {}", error),
+                    )
+                },
+            ),
+        )
+        .route(
+            "/robots.txt",
+            get_service(ServeFile::new("./public/robots.txt")).handle_error(
                 |error: std::io::Error| async move {
                     (
                         StatusCode::INTERNAL_SERVER_ERROR,
@@ -63,8 +56,20 @@ async fn main() {
         .route("/", get(blog_index_handler))
         .route("/blog/:blog", get(blog_handler))
         .route("/category/:category", get(category_handler))
-        .nest("/series", series_routes)
-        .route("/about", get(about_handler));
+        .route("/series", get(series_index_handler))
+        .route("/series/:series", get(series_handler))
+        .route("/about", get(about_handler))
+        .nest(
+            "/public",
+            get_service(ServeDir::new("./public/")).handle_error(
+                |error: std::io::Error| async move {
+                    (
+                        StatusCode::INTERNAL_SERVER_ERROR,
+                        format!("Unhandled internal error: {}", error),
+                    )
+                },
+            ),
+        );
 
     let addr = SocketAddr::from(([0, 0, 0, 0], 8000));
     println!("Server: {}", addr);
@@ -72,6 +77,14 @@ async fn main() {
         .serve(app.into_make_service())
         .await
         .unwrap();
+}
+
+#[derive(Template)]
+#[template(path = "not_found.html")]
+struct NotFoundTemplate {}
+async fn handler_404() -> impl IntoResponse {
+    let template = NotFoundTemplate {};
+    HtmlTemplate(template)
 }
 
 // Not ready for production