+++ CHANGELOG.md
@@ -7,6 +7,22 @@ Addition of markdown posts will not be versioned or require a changelog update.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2022-05-06
+
+### Added
+
+- handler_404 template [wip]
+
+### Changed
+
+- Public assets now route to `./public`
+- sw.js and robots.txt are still served on `/`
+- Fallback now works correctly with tower-http 3^
+
+### Removed
+
+- tower-http fallback route
+
 ## [0.7.2] - 2022-05-06
 ### - Will not compile